==> AF_MA_Master_LAA/13_Added_value.Rmd <==
---
title: "13_Added_value"
output:
  word_document: default
  html_document: default
date: "2023-02-26"
---

```{r}
library(PLIER)
library(reshape2)
library(ROCR)
library(fgsea)
library(dplyr)
library(tidyr)
library(WriteXLS)

source("src/utils_AF_Meta.R")
```

# 2. Main data object importing
```{r}
AF_Meta = readRDS(file = "data/AF_Meta.rds") #main object

AF_Meta_RAA = readRDS(file = "data/AF_Meta_RAA.rds") #main object
```

```{r}
experiments = names(AF_Meta)
names(experiments) = experiments

experiments_RAA = names(AF_Meta_RAA)
names(experiments_RAA) = experiments_RAA
```

```{r}
set.seed(12345)

# For labeling
t_matrix = get_all_limma_AF(meta_list = AF_Meta,
                         limma_column = "t")

# For labeling
t_matrix_RAA = get_all_limma_AF(meta_list = AF_Meta_RAA,
                         limma_column = "t")

# Here we get AUC for pairwise classifiers
pairwise_500 = pairwise_ds_AF(experiments = experiments,
                           meta_list = AF_Meta,
                           t_matrix = t_matrix,
                           ngenes = 500) #Second page excel

# Here we get AUC for pairwise classifiers
pairwise_500_RAA = pairwise_ds_AF(experiments = experiments_RAA,
                           meta_list = AF_Meta_RAA,
                           t_matrix = t_matrix_RAA,
                           ngenes = 500) #Second page excel
```

```{r}
# Here we get the results from performing the meta-analysis

fisher_rank = run_fisher_meta_AF(meta_list = AF_Meta,
                              n_missing = length(AF_Meta) - 5)

sum(fisher_rank < .00005)

genes = names(fisher_rank)

fisher_rank_RAA = run_fisher_meta_AF(meta_list = AF_Meta_RAA,
                              n_missing = length(AF_Meta_RAA) - 5)
```

```{r}
# Here we get the calculations of using the top N genes from the meta-analysis

ds_top = getRisk_Stats_AF(Experiment_List = AF_Meta,
                          limma_t_mat = t_matrix, 
                          genes = names(fisher_rank[1:500]))

ds_top_predictions = enframe(lapply(ds_top, function(x) {
  enframe(x[["SingleAUC"]])
})) %>% unnest()

colnames(ds_top_predictions) = c("PredictedExperiment", 
                                 "PredictorExperiment", 
                                 "meta_auc")

ds_top_RAA_alone = getRisk_Stats_AF(Experiment_List = AF_Meta_RAA,
                          limma_t_mat = t_matrix_RAA, 
                          genes = names(fisher_rank_RAA[1:500]))

ds_top_predictions_RAA_alone = enframe(lapply(ds_top_RAA_alone, function(x) {
  enframe(x[["SingleAUC"]])
})) %>% unnest()

colnames(ds_top_predictions_RAA_alone) = c("PredictedExperiment", 
                                 "PredictorExperiment", 
                                 "meta_auc")



ds_top_RAA = getRisk_Stats_AF(Experiment_List = AF_Meta,
                          limma_t_mat = t_matrix, 
                          genes = names(fisher_rank_RAA[1:500]))

ds_top_predictions_RAA = enframe(lapply(ds_top_RAA, function(x) {
  enframe(x[["SingleAUC"]])
})) %>% unnest()

colnames(ds_top_predictions_RAA) = c("PredictedExperiment", 
                                 "PredictorExperiment", 
                                 "meta_auc")


ds_top_LAA = getRisk_Stats_AF(Experiment_List = AF_Meta_RAA,
                          limma_t_mat = t_matrix_RAA, 
                          genes = names(fisher_rank[1:500]))

ds_top_predictions_LAA = enframe(lapply(ds_top_LAA, function(x) {
  enframe(x[["SingleAUC"]])
})) %>% unnest()

colnames(ds_top_predictions_LAA) = c("PredictedExperiment", 
                                 "PredictorExperiment", 
                                 "meta_auc")
```

```{r}
# Here we merge them 

comp_df = left_join(pairwise_500,
                    ds_top_predictions) %>%
          dplyr::filter(PredictedExperiment != PredictorExperiment) %>%
  dplyr::select(PredictorExperiment, 
                PredictedExperiment,
                single, meta_auc)

print("Are AUCs better?")

wilcox.test(comp_df$meta_auc,comp_df$single,paired = T,alternative = "greater")

comp_df = comp_df %>%
  gather("type","AUROC",
         -PredictorExperiment,
         -PredictedExperiment)

saveRDS(comp_df, "data/figure_objects/added_value_ds.rds")

comp_df_RAA = left_join(pairwise_500_RAA,
                    ds_top_predictions_RAA_alone) %>%
          dplyr::filter(PredictedExperiment != PredictorExperiment) %>%
  dplyr::select(PredictorExperiment, 
                PredictedExperiment,
                single, meta_auc)

print("Are AUCs better?")

wilcox.test(comp_df_RAA$meta_auc,comp_df_RAA$single,paired = T,alternative = "greater")


comp_df_LAA_with_RAA_top_meta = left_join(pairwise_500,
                    ds_top_predictions_RAA) %>%
          dplyr::filter(PredictedExperiment != PredictorExperiment) %>%
  dplyr::select(PredictorExperiment, 
                PredictedExperiment,
                single, meta_auc)

print("Are AUCs of the RAA MA better than the individual LAA studies?")

wilcox.test(comp_df_LAA_with_RAA_top_meta$meta_auc,comp_df_LAA_with_RAA_top_meta$single,paired = T,alternative = "greater")


comp_df_RAA_with_LAA_top_meta = left_join(pairwise_500_RAA,
                    ds_top_predictions_LAA) %>%
          dplyr::filter(PredictedExperiment != PredictorExperiment) %>%
  dplyr::select(PredictorExperiment, 
                PredictedExperiment,
                single, meta_auc)

print("Are AUCs of the LAA MA better than the individual RAA studies?")

wilcox.test(comp_df_RAA_with_LAA_top_meta$meta_auc,comp_df_RAA_with_LAA_top_meta$single,paired = T,alternative = "greater")
```

```{r}
# Here we test the added value at consistency

study_deg_list_up = lapply(AF_Meta, function(x){
  deg = dplyr::filter(x$AF_limma,
                     ID %in% names(fisher_rank[1:500])) %>%
    filter(t >0)
  return(deg[[1]])
})

study_deg_list_down = lapply(AF_Meta, function(x){
  deg = dplyr::filter(x$AF_limma,
                      ID %in% names(fisher_rank[1:500])) %>%
    filter(t < 0)
  return(deg[[1]])
})

# upregulation

up_ES = lapply(experiments, function(x){
  
  stat_rank = AF_Meta[[x]][["AF_limma"]][["t"]]
  names(stat_rank) = AF_Meta[[x]][["AF_limma"]][["ID"]]
  stat_rank = sort(stat_rank)
  set.seed(1234)
  
  up_row = as_tibble(fgsea(pathways = study_deg_list_up,
                           stats = stat_rank,nperm = 1000)) %>%
    dplyr::select(pathway,ES)
})

up_ES = up_ES %>% 
  enframe("Reference") %>% unnest()

colnames(up_ES) = c("Reference","DEG","ES")
```

```{r}
# downregulation

down_ES = lapply(experiments, function(x){
  
  stat_rank = AF_Meta[[x]][["AF_limma"]][["t"]]
  names(stat_rank) = AF_Meta[[x]][["AF_limma"]][["ID"]]
  stat_rank = sort(stat_rank)
  set.seed(1234)
  
  up_row = as_tibble(fgsea(pathways = study_deg_list_down,
                           stats = stat_rank,nperm = 1000)) %>%
    dplyr::select(pathway,ES)
})

down_ES = down_ES %>% 
  enframe("Reference") %>% unnest()

colnames(down_ES) = c("Reference","DEG","ES")

# Previous results
up_ES_single = readRDS(file = "data/figure_objects/up_ES.rds")
down_ES_single = readRDS(file = "data/figure_objects/down_ES.rds")


# Merge them
print("Is the ES better?")

comp_ES_up= left_join(up_ES_single,
                      up_ES, by = c("Reference","DEG")) %>%
  dplyr::filter(Reference != DEG) 

wilcox.test(comp_ES_up$ES.y,
            comp_ES_up$ES.x,
            paired = T,
            alternative = "greater")

comp_ES_down = left_join(down_ES_single,
                      down_ES, by = c("Reference","DEG")) %>%
  dplyr::filter(Reference != DEG)

wilcox.test(comp_ES_down$ES.y,
            comp_ES_down$ES.x,
            paired = T,
            alternative = "less")

comp_ES = bind_rows("upregulated"=comp_ES_up, 
                    "downregulated" =comp_ES_down,
                    .id = "comparison") %>% 
  mutate(single = ES.x, 
         meta_analysis = ES.y) %>%
  dplyr::select(c("comparison","Reference","DEG",
                  "single","meta_analysis")) %>%
  gather("type","ES", 
         -comparison,-Reference, 
         -DEG)

saveRDS(comp_ES, "data/figure_objects/added_value_es.rds")

```
