==> AF_MA_Master_LAA/14_gene_variability.Rmd <==
---
title: "14_gene_variability"
output: html_document
date: "2023-02-26"
---

```{r}
source("src/utils_AF_Meta.R")
AF_Meta = readRDS(file = "data/AF_Meta.rds") #main object
```

```{r}
AF_Meta = lapply(AF_Meta, function(x){
  x[["GEX_norm"]] = PLIER::rowNorm(x[["GEX"]])
  return(x)
})
```


```{r}
# Generating a unified data set

meta_targets = get_tibble_union(AF_Meta,"TARGETS") %>% 
  dplyr::select(sample,condition,ExpID) %>% 
  mutate(grl_id = paste(sample,ExpID,sep = "_"))

meta_gex = get_complete_gex_AF(meta_list = AF_Meta,
                            complete_targets = meta_targets,
                            gex_key = "GEX")

meta_gex_scale = get_complete_gex_AF(meta_list = AF_Meta,
                                  complete_targets = meta_targets,
                                  gex_key = "GEX_norm")
```

```{r}
meta_targets = meta_targets %>% mutate(Tech = case_when(ExpID == "Hsu_2018" | ExpID == "Thomas_2019" | ExpID == "Sun_2021" | ExpID == "Herrera_Rivero_2022" | ExpID == "Santos_2020" | ExpID == "Zhu_2020"  ~ "RNA-Seq", TRUE ~ "microarray"))
```

```{r}
# Modelling each factor

fisher_rank = run_fisher_meta_AF(meta_list = AF_Meta ,
                              n_missing = length(AF_Meta) - 5)

genes = names(fisher_rank)

```

```{r}
# Performing a 2-way anova with interaction term (AF/study)

anova_study_scale = run_anovastats_AF(numeric_matrix = meta_gex_scale[genes,],
                                   targets = meta_targets) #For sup

```

```{r}
anova_confounding = list("study" = anova_study_scale)

anova_confounding = map(anova_confounding, function(x){
  
  mutate(x,data_type = "gene_centered") %>%
    dplyr::filter(stats=="etasq")
  
} )

saveRDS(anova_confounding, 
        file = "data/figure_objects/confounding_anova.rds")
```

```{r}
library(tidyverse)
library(cowplot)
```

```{r}
anova_confounding = readRDS("data/figure_objects/confounding_anova.rds")

fisher_rank = run_fisher_meta_AF(meta_list = AF_Meta,
                              n_missing = 4)

genes = names(fisher_rank)

ngenes = 450
```

```{r}
# Data frame with ANOVA results

anova_confounding_mod = enframe(lapply(anova_confounding, 
                                       function(x){
                                         colnames(x)[5] = "cfactor"
                                         return(x)
                                       }), "cfactor_name", "values") %>% unnest() %>% 
  dplyr::filter(data_type == "gene_centered") %>% 
  filter(stats=="etasq") %>% 
  mutate(significant = ifelse(gene %in% genes[1:ngenes],
                              "yes","no")) %>% 
  arrange(significant)
```

```{r}
# Upper panel

anova_plt_v2 = ggplot(anova_confounding_mod,
                      aes(x=condition, 
                          y = cfactor, 
                          color = factor(significant,levels = c("SR","AF")))) + 
  geom_point(alpha = 2/3) + 
  facet_grid(.~factor(cfactor_name,
                      levels = c("study"
                                 ))) +
  theme_minimal() + scale_color_manual(values=c("lightgrey", "black")) +
  theme(legend.position = "none",
        panel.grid = element_blank(),
        axis.text.y=element_text(size=10),
        axis.text.x = element_text(size=10),
        axis.title.x = element_text(size=12),
        axis.title.y = element_text(size=12),
        strip.text.x = element_text(size=12),
        panel.background = element_rect(fill=NULL, colour='black',
                                        size=1)) + xlab("eta squared Atrial Fibrillation") +
  ylab("eta squared factor")
```

```{r}
# Lower panel

anova_boxplt_v2 = ggplot(anova_confounding_mod,
                         aes(y = condition, 
                             x = factor(significant,levels = c("AF",
                                                               "SR")))) + 
  geom_boxplot() + 
  facet_grid(.~factor(cfactor_name,
                      levels = c("study"))) +
  theme_minimal() + 
  theme(legend.position = "none",
        panel.grid = element_blank(),
        axis.text.y=element_text(size=10),
        axis.text.x = element_text(size=10),
        axis.title.x = element_text(size=12),
        axis.title.y = element_text(size=12),
        strip.text.x = element_text(size=12),
        panel.background = element_rect(fill=NULL, 
                                        colour='black',
                                        size=1)) + 
  theme(strip.text.x = element_blank()) +
  xlab("Genes in top 450") +
  ylab("Prop of variance explained by AF")
```

```{r}
# Create figure
pdf("data/figures/sup/SupplementalFigure11.pdf",
    width = 10,
    height = 8)

plot(plot_grid(anova_plt_v2, anova_boxplt_v2,
               align = "h", ncol = 1,
               rel_heights = c(.5,1)))

dev.off()
```
